Moved navbar to separate file; added props validation
